whisper-server/server.py
#!/usr/bin/env python3
"""
Local Whisper Server - Runs OpenAI Whisper on your GPU
Provides REST API for audio transcription
"""

import os
import sys
import torch
import whisper
import numpy as np
import soundfile as sf
import io
import json
import logging
import time
from pathlib import Path
from flask import Flask, request, jsonify
from flask_cors import CORS
from werkzeug.utils import secure_filename
import warnings

# Suppress warnings
warnings.filterwarnings("ignore")

# Setup logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

# Initialize Flask app
app = Flask(__name__)
CORS(app)  # Enable CORS for all routes

# Configuration
MAX_FILE_SIZE = 25 * 1024 * 1024  # 25MB
ALLOWED_EXTENSIONS = {'wav', 'mp3', 'mp4', 'm4a', 'ogg', 'flac', 'webm'}
MODEL_SIZE = "base"  # Options: tiny, base, small, medium, large

# Global variables
model = None
device = None

def init_whisper():
    """Initialize Whisper model with GPU support"""
    global model, device
    
    # Check for CUDA availability
    if torch.cuda.is_available():
        device = "cuda"
        logger.info(f"🎉 GPU detected: {torch.cuda.get_device_name(0)}")
        logger.info(f"   VRAM: {torch.cuda.get_device_properties(0).total_memory / 1024**3:.1f} GB")
    else:
        device = "cpu"
        logger.warning("⚠️  No GPU detected, using CPU (will be slower)")
    
    # Load Whisper model
    logger.info(f"📥 Loading Whisper model: {MODEL_SIZE}")
    start_time = time.time()
    
    try:
        model = whisper.load_model(MODEL_SIZE, device=device)
        load_time = time.time() - start_time
        logger.info(f"✅ Model loaded successfully in {load_time:.1f} seconds")
    except Exception as e:
        logger.error(f"❌ Failed to load model: {str(e)}")
        sys.exit(1)

def allowed_file(filename):
    """Check if file extension is allowed"""
    return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS

@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({
        "status": "healthy",
        "model": MODEL_SIZE,
        "device": device,
        "gpu_available": torch.cuda.is_available(),
        "gpu_name": torch.cuda.get_device_name(0) if torch.cuda.is_available() else None
    })

@app.route('/api/transcribe', methods=['POST'])
def transcribe_audio():
    """Main transcription endpoint"""
    start_time = time.time()
    
    # Check if file was uploaded
    if 'file' not in request.files:
        return jsonify({"error": "No file provided"}), 400
    
    file = request.files['file']
    if file.filename == '':
        return jsonify({"error": "No file selected"}), 400
    
    if not allowed_file(file.filename):
        return jsonify({"error": f"Invalid file type. Allowed: {', '.join(ALLOWED_EXTENSIONS)}"}), 400
    
    try:
        # Log file info
        file_size = len(file.read())
        file.seek(0)  # Reset file pointer
        logger.info(f"📁 Received file: {file.filename} ({file_size / 1024 / 1024:.1f} MB)")
        
        # Check file size
        if file_size > MAX_FILE_SIZE:
            return jsonify({"error": f"File too large. Max size: {MAX_FILE_SIZE / 1024 / 1024} MB"}), 400
        
        # Save file temporarily
        temp_path = Path("temp_audio") / secure_filename(file.filename)
        temp_path.parent.mkdir(exist_ok=True)
        file.save(temp_path)
        
        # Transcribe with Whisper
        logger.info("🎙️ Starting transcription...")
        
        # Run transcription
        result = model.transcribe(
            str(temp_path),
            language="en",  # Force English (remove for auto-detect)
            task="transcribe",
            fp16=torch.cuda.is_available(),  # Use FP16 on GPU
            verbose=False
        )
        
        # Clean up temp file
        temp_path.unlink()
        
        # Calculate metrics
        process_time = time.time() - start_time
        logger.info(f"✅ Transcription completed in {process_time:.1f} seconds")
        
        # Extract segments with timestamps
        segments = []
        for segment in result.get("segments", []):
            segments.append({
                "start": segment["start"],
                "end": segment["end"],
                "text": segment["text"].strip()
            })
        
        # Return response
        return jsonify({
            "transcript": result["text"].strip(),
            "confidence": 0.95,  # Whisper doesn't provide confidence scores
            "duration": process_time,
            "model": f"whisper-{MODEL_SIZE}",
            "device": device,
            "segments": segments,
            "language": result.get("language", "en"),
            "processing_stats": {
                "file_size_mb": round(file_size / 1024 / 1024, 2),
                "processing_time_seconds": round(process_time, 2),
                "speed_factor": round(result.get("segments", [])[-1]["end"] / process_time if segments else 0, 1)
            }
        })
        
    except Exception as e:
        logger.error(f"❌ Transcription error: {str(e)}")
        return jsonify({
            "error": "Transcription failed",
            "details": str(e)
        }), 500

@app.route('/api/models', methods=['GET'])
def list_models():
    """List available Whisper models"""
    models = {
        "tiny": {"size": "39M", "speed": "~32x", "quality": "Good for quick drafts"},
        "base": {"size": "74M", "speed": "~16x", "quality": "Good balance (recommended)"},
        "small": {"size": "244M", "speed": "~6x", "quality": "Better accuracy"},
        "medium": {"size": "769M", "speed": "~2x", "quality": "High accuracy"},
        "large": {"size": "1550M", "speed": "~1x", "quality": "Best accuracy"}
    }
    
    return jsonify({
        "current_model": MODEL_SIZE,
        "available_models": models,
        "gpu_supported": torch.cuda.is_available(),
        "recommendation": "Use 'base' for RTX 4050 6GB - best balance of speed and quality"
    })

@app.route('/api/test', methods=['GET'])
def test_endpoint():
    """Test endpoint to verify server is running"""
    return jsonify({
        "message": "Whisper server is running!",
        "model": f"whisper-{MODEL_SIZE}",
        "device": device,
        "timestamp": time.time()
    })

if __name__ == '__main__':
    print("""
    ╔══════════════════════════════════════════╗
    ║     🎙️  Local Whisper Server v1.0       ║
    ╚══════════════════════════════════════════╝
    """)
    
    # Initialize Whisper
    init_whisper()
    
    print(f"""
    🚀 Server Configuration:
       Model: whisper-{MODEL_SIZE}
       Device: {device}
       Max file size: {MAX_FILE_SIZE / 1024 / 1024} MB
    
    📡 Starting server on http://localhost:5000
    
    🔗 Endpoints:
       POST /api/transcribe - Transcribe audio
       GET  /api/models     - List available models
       GET  /health         - Health check
       GET  /api/test       - Test endpoint
    
    💡 Tips:
       - First transcription will download the model (~74MB)
       - Subsequent transcriptions will be much faster
       - Your RTX 4050 will process audio 10-20x faster than real-time
    """)
    
    # Run Flask server
    app.run(host='0.0.0.0', port=5000, debug=False)
