whisper-server/requirements.txt
openai-whisper==20231117
torch>=2.0.0
flask==3.0.0
flask-cors==4.0.0
soundfile==0.12.1
numpy==1.24.4
werkzeug==3.0.1
